Move DreamIcon out of DreamMetaObjectIcon
A class this large and with this much functionality should probably be its own thing
